Troy-backend/requirements.txt
aniso8601==9.0.1
atomicwrites==1.4.1
attrs==22.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
deep-translator==1.11.4
docx2txt==0.8
et-xmlfile==1.1.0
filelock==3.12.2
Flask==2.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.1
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
fsspec==2023.6.0
googletrans==4.0.0rc1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.16.2
hyperframe==5.2.0
idna==2.10
importlib-metadata==5.1.0
importlib-resources==5.10.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
langid==1.1.6
lxml==4.9.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.24.4
openpyxl==3.1.2
packaging==22.0
pandas==2.0.3
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
py==1.11.0
PyJWT==2.6.0
pyrsistent==0.19.2
pytest==7.1.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.2
pytz==2022.6
PyYAML==6.0
regex==2023.6.3
requests==2.28.1
rfc3986==1.5.0
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.46
sympy==1.12
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.13
Werkzeug==2.1.2
zipp==3.11.0
